+++ packages/api-post/src/groupPosts.ts
@@ -19,12 +19,14 @@ export const fetchGroupPosts = ({
   offset = 0,
   limit = 50,
   tags = [],
+  users = [],
   sort = Sorting.latest,
 }: {
   group: string
   offset?: number
   limit?: number
   tags?: string[]
+  users?: string[]
   sort?: Sorting
 }) => {
   const params = new URLSearchParams()
@@ -35,6 +37,12 @@ export const fetchGroupPosts = ({
   if (tags.length > 0) {
     for (const tag of tags) {
       params.append('tags[]', tag)
+    }
+  }
+
+  if (users.length > 0) {
+    for (const user of users) {
+      params.append('users[]', user)
     }
   }